Top data in infobox is a list item now
