+++ src/Forvaltningsportalen/FeatureInfo/ClickInfobox.js
@@ -1,7 +1,8 @@
 import React, { useState, useCallback } from "react";
 import { Place, Home, Flag, Terrain } from "@material-ui/icons";
+import { ListItem } from "@material-ui/core";
 import CustomTooltip from "../../Common/CustomTooltip";
-import CustomSwitch from "../../Common/CustomSwitch";
+// import CustomSwitch from "../../Common/CustomSwitch";
 import "../../style/infobox.css";
 import ResultsList from "./ResultsList";
 import DetailedResults from "./DetailedResults";
@@ -55,12 +56,6 @@ const ClickInfobox = ({
     return matrikkel;
   };
 
-  const toggleAllLayers = () => {
-    if (coordinates_area && coordinates_area.lng && coordinates_area.lat) {
-      getBackendData(coordinates_area.lng, coordinates_area.lat);
-    }
-  };
-
   const showDetailedResults = useCallback(
     (layer, listResults, primaryText, secondaryText, numberResults) => {
       // Remember scroll position of infobox
@@ -86,7 +81,6 @@ const ClickInfobox = ({
   );
 
   const hideDetailedResults = () => {
-    setInfoboxDetailsVisible(false);
     setResultLayer(null);
     setListResults(null);
     setPrimaryText(null);
@@ -97,6 +91,7 @@ const ClickInfobox = ({
     // Set scroll position to original value
     let wrapper = document.querySelector(".infobox-side");
     setTimeout(() => {
+      setInfoboxDetailsVisible(false);
       wrapper.scrollTop = infoboxScroll;
     }, 5);
   };
@@ -116,63 +111,72 @@ const ClickInfobox = ({
         />
       )}
       <div className={infoboxDetailsVisible ? "infobox-content-hidden" : ""}>
-        <div className="infobox-content">
-          <div className="infobox-text-wrapper">
-            <CustomTooltip placement="right" title="Fylke / Fylkesnr.">
-              <Terrain />
-            </CustomTooltip>
-            <div className="infobox-text-multiple">
-              <div className="infobox-text-primary">
-                {sted ? sted.fylkesnavn[0] : "-"}
+        <ListItem
+          id="infobox-main-content-button"
+          button
+          onClick={e => {
+            console.log("Open");
+          }}
+        >
+          <div className="infobox-content">
+            <div className="infobox-text-wrapper">
+              <CustomTooltip placement="right" title="Fylke / Fylkesnr.">
+                <Terrain />
+              </CustomTooltip>
+              <div className="infobox-text-multiple">
+                <div className="infobox-text-primary">
+                  {sted ? sted.fylkesnavn[0] : "-"}
+                </div>
+                <div className="infobox-text-secondary">
+                  {sted ? sted.fylkesnummer[0] : "-"}
+                </div>
               </div>
-              <div className="infobox-text-secondary">
-                {sted ? sted.fylkesnummer[0] : "-"}
+            </div>
+            <div className="infobox-text-wrapper">
+              <CustomTooltip placement="right" title="Kommune / Kommunenr.">
+                <Flag />
+              </CustomTooltip>
+              <div className="infobox-text-multiple">
+                <div className="infobox-text-primary">
+                  {sted ? sted.kommunenavn[0] : "-"}
+                </div>
+                <div className="infobox-text-secondary">
+                  {sted ? sted.kommunenummer[0] : "-"}
+                </div>
+              </div>
+            </div>
+            <div className="infobox-text-wrapper">
+              <CustomTooltip
+                placement="right"
+                title="Adresse / Gårdsnr. / Bruksnr."
+              >
+                <Home />
+              </CustomTooltip>
+              <div className="infobox-text-multiple">
+                <div className="infobox-text-primary">
+                  {hentAdresse(adresse)}
+                </div>
+                <div className="infobox-text-secondary">
+                  {hentMatrikkel(matrikkel)}
+                </div>
+              </div>
+            </div>
+            <div className="infobox-text-wrapper">
+              <CustomTooltip placement="right" title="Koordinater / Høyde">
+                <Place />
+              </CustomTooltip>
+              <div className="infobox-text-multiple">
+                <div className="infobox-text-primary">
+                  {coordinates_area ? coords : "--° N --° Ø"}
+                </div>
+                <div className="infobox-text-tertyary">
+                  {elevation ? elevation + " moh" : "-"}
+                </div>
               </div>
             </div>
           </div>
-          <div className="infobox-text-wrapper">
-            <CustomTooltip placement="right" title="Kommune / Kommunenr.">
-              <Flag />
-            </CustomTooltip>
-            <div className="infobox-text-multiple">
-              <div className="infobox-text-primary">
-                {sted ? sted.kommunenavn[0] : "-"}
-              </div>
-              <div className="infobox-text-secondary">
-                {sted ? sted.kommunenummer[0] : "-"}
-              </div>
-            </div>
-          </div>
-          <div className="infobox-text-wrapper">
-            <CustomTooltip
-              placement="right"
-              title="Adresse / Gårdsnr. / Bruksnr."
-            >
-              <Home />
-            </CustomTooltip>
-            <div className="infobox-text-multiple">
-              <div className="infobox-text-primary">{hentAdresse(adresse)}</div>
-              <div className="infobox-text-secondary">
-                {/* {`${hentGardsnummer(adresse)}/${hentBruksnummer(adresse)}`} */}
-                {hentMatrikkel(matrikkel)}
-              </div>
-            </div>
-          </div>
-          <div className="infobox-text-wrapper">
-            <CustomTooltip placement="right" title="Koordinater / Høyde">
-              <Place />
-            </CustomTooltip>
-            <div className="infobox-text-multiple">
-              <div className="infobox-text-primary">
-                {coordinates_area ? coords : "--° N --° Ø"}
-              </div>
-              <div className="infobox-text-tertyary">
-                {elevation ? elevation + " moh" : "-"}
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="show-property-button-wrapper">
+        </ListItem>
+        {/* <div className="show-property-button-wrapper">
           <span className="infobox-switch-text">Gjem eiendom</span>
           <CustomSwitch
             tabIndex="0"
@@ -186,24 +190,7 @@ const ClickInfobox = ({
             }}
           />
           <span className="infobox-switch-text">Vis eiendom</span>
-        </div>
-        <div className="search-layers-button-wrapper">
-          <span className="infobox-switch-text" style={{ paddingRight: "4px" }}>
-            Valgte kartlag
-          </span>
-          <CustomSwitch
-            tabIndex="0"
-            id="search-layers-toggle"
-            checked={showExtensiveInfo}
-            onChange={toggleAllLayers}
-            onKeyDown={e => {
-              if (e.keyCode === 13) {
-                toggleAllLayers();
-              }
-            }}
-          />
-          <span className="infobox-switch-text">Alle kartlag</span>
-        </div>
+        </div> */}
         <ResultsList
           showExtensiveInfo={showExtensiveInfo}
           kartlag={showExtensiveInfo ? kartlag : valgteLag}
@@ -217,6 +204,7 @@ const ClickInfobox = ({
           sortKey={sortKey}
           tagFilter={tagFilter}
           matchAllFilters={matchAllFilters}
+          getBackendData={getBackendData}
         />
       </div>
     </div>
